drm/i915: Expose a context ID via i915 perf samples
TODO: document as a whole...

drm/i915: Introduce mode for capture of CS based OA reports

This patch introduces a mode of capturing OA counter reports belonging to
multiple contexts, which can be mapped back to individual contexts. The OA
reports captured in this way are synchronized with Render command stream.

There may be usecases wherein we need more than periodic OA capture mode
which is supported currently. We may need to insert RCS synchronized
commands to capture the OA counter snapshots.
This mode is primarily used for two usecases:
    - Ability to capture system wide metrics, alongwith the ability to map
      the reports back to individual contexts.
    - Ability to inject tags for work, into the reports. This provides
      visibility into the multiple stages of work within single context.

The OA reports generated in this way will be forwarded to userspace
alongwith the metadata information. This will enable the usecases mentioned
above.

The data thus captured needs to be stored in a separate buffer, which will
be different from the buffer used otherwise for periodic OA capture mode.
Again this buffer address will not need to be mapped to OA unit register
addresses such as OASTATUS1, OASTATUS2 and OABUFFER.

The userspace will be able to distinguish between the periodic and CS based
OA reports by the virtue of a 'source_info' sample field, which is enabled
by _SOURCE_INFO sample flag.

The subsequent patches introduce the  mechanism for forwarding reports to
userspace, handling the command synchronization and mechanism for inserting
corresponding commands into the ringbuffer.

drm/i915: Add mechanism for reading CS based OA snapshots

This patch adds the mechanism for forwarding the CS based OA snapshots,
through a user initiated read call.

The OA snapshots will be captured in a gem buffer object. The metadata
information pertaining to snapshot is maintained in a list, which also has
offsets into the gem buffer object for each snapshot captured.
Each snapshot collected is forwarded as a separate sample, alongwith its
metadata info. The size of the OA report is the one specified during event
init.

In order to track whether the gpu has completed processing the node,
a field pertaining to corresponding gem request is added. The request is
expected to be referenced whenever the gpu command is submitted.

While forwarding the samples, we check whether the gem request is completed
and dereference the corresponding request.
While the event is being stopped, we need to wait for gpu to complete
processing the submitted requests before dereferencing those requests,
and disabling the OA unit.

drm/i915: Sort periodic and CS based OA reports acc to timestamps

The periodic reports and the RCS based reports are collected in separate
buffers, whereas there is a single read call to collect both these buffer.
Its imperative to have these OA report sorted based on timestamps before
forwarding to userspace.
This patch addresses this specific problem by merge sorting these two,
before copying to userspace.

drm/i915: Insert commands for capture of OA counters in the ring

This patch adds the routines which insert commands for capturing OA reports
into the ringbuffer of RCS engine.

The command MI_REPORT_PERF_COUNT can be used to capture snapshots of OA
counters, which is inserted at BB boundaries.
While inserting the commands, we keep a reference of associated request.
This will be released when we are forwarding the samples to userspace
(or when the event is being destroyed).
Also, an active reference of the destination buffer is taken here, so that
we can be assured that the buffer is freed up only after GPU is done with
it, even if the local reference of the buffer is released.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>
Signed-off-by: Robert Bragg <robert@sixbynine.org>