drm/i915: Extend i915 perf framework to all engines for collecting timestamps
This patch extends the i915  perf framework to handle the perf sample
collection for multiple gpu engines (rings). Particularly, the support
for collecting timestamp sample type is added, which can be requested for
any ring.
This is done by opening a separate stream associated with the corresponding
engine to collect the samples. Multiple such streams can be opened concurrently,
each tied to different gpu engine.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>