drm/i915: Add an I915_PERF_TIMSTAMP_EVENT
TODO: document as a whole..
TODO: Rename _GEN_EVENT to _TIMESTAMP_EVENT. We can plan to add other
      more specific events e.g. for capturing mmio registers too.

      Notably this departs a bit further from the core perf pmu based
      model but that model doesn't seem to buy us anything here.
TODO: See if this can be more integrated with the oa rcs mode support

drm/i915: Framework for handling non-OA profiling data

The current i915 perf infrastructure is specific for collection of OA
counter reports. Since this enables us (and limits us) to the render ring,
we have no means for collection of profiling data pertaining to other rings

To overcome this limitation, we need to have a way and means of capturing
performance data for other rings also (in a non-OA specific way).
This data may encompass timestamps, mmio register values, etc.

This new infrastructure will serve these purposes, without constraining
itself to type of data being dumped (which may restrict the user to
specific ring like in case of OA counters).

The patch introduces this framework and its associated callbacks.
The subsequent patches build on this capability to capture timestamps,
mmio metrics etc. via this framework.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>

drm/i915: Add mechanism for forwarding the timestamp data

This patch adds the mechanism for forwarding the timestamp data to
userspace, through a user initiated read call.

The timestamps will be captured in a gem buffer object. The metada
information pertaining to snapshot is maintained in a list, whose each node
has offsets into the gem buffer object for each snapshot captured.
In order to track whether the gpu has completed processing the node,
a field pertaining to corresponding gem request is added. The request is
expected to be referenced whenever the gpu command is submitted.

While forwarding the samples, we check whether the gem request is completed
and dereference the corresponding request.
While the event is being stopped, we need to wait for gpu to complete
processing the submitted requests before dereferencing those requests.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>

drm/i915: Insert commands for capturing timestamps in the ring

This patch adds the routines through which one can insert commands in the
ringbuf for capturing timestamps, around the batchbuffer.

While inserting the commands, we keep a reference of associated request.
This will be released when we are forwarding the samples to userspace
(or when the event is being destroyed).
Also, an active reference of the destination buffer is taken here,
so that we can be assured that the buffer is freed up only after GPU is
done with it, even if the local reference of the buffer is released.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>