drm/i915: start tracking device_info subslice_mask
Not assuming the subslice mask is uniform across all slices this mask
covers all slices with N bits per slice, where N is the maximum number
of slices for a particular Gen.

Notably this matches the definition of the $SubsliceMask variable that
is currently found in out OA metrics XML files.

Cc: Zhenyu Wang <zhenyuw@linux.intel.com>
Signed-off-by: Robert Bragg <robert@sixbynine.org>