drm/i915: Add support for forwarding ring id in the sample
This patch introduces flags and adds support for having ring id output with
the timestamp samples and forwarding them to userspace.

When the userspace expresses its interest in listening to the ring id
through the corresponding sample flags, the samples generated would have an
additional field with the ring id information.

Signed-off-by: Sourab Gupta <sourab.gupta@intel.com>