stash: rename OA_EVENT -> RING_EVENT
This generalizes the I915_PERF_OA_EVENT into a _RING_EVENT which is
configured for a given ring ID and OA metrics are conceptually just one
of many things that could be sampled with respect to the given ring.

XXX: There is a notable implication when requesting OA samples with
respect to a ring which userspace needs to be aware of. Later when we
actually start capturing OA metrics via MI_RPC commands (or other
metrics via the CS) then periodic OA samples will be merge sorted with
the CS captured metrics. This implies periodic samples are throttled by
the receipt of CS metrics which userspace should be aware of because if
they aren't actually issuing any commands they won't receive any
periodic samples.

Given this we should consider whether it makes sense to also support an
_OA_EVENT like before that's unrelated to a ring in addition to this
_RING_EVENT + OA sampling. If we supported both kinds then notably we
would only support _SAMPLE_CTXID on Haswell for _RING_EVENT.