drm/i915: Implement I915_PERF_ADD_CONFIG interface
The motivation behind this new interface is expose at runtime
the creation of new OA configs which can be used as part of the
i915 perf open interface. This will enable the kernel to learn new
configs which may be experimental, or otherwise not part of the core set
currently available through the i915 perf interface.